Prevent duplicate collection and always run renaming/label logic
This change prevents running the replSetGetStatus or serverStatus
collectors if those data have been collected already via
getDiagnosticData.

Provide data from the replSetGetStatus and serverStatus collectors to
makeMetrics in a format such that the desired renames and label
management actions are carried out.

Note that this change is not backwards compatible for anyone running
with --collector.replicasetstatus as those metrics will now have a
prefix of mongodb_rs_ rather than just mongodb_